+++ plugins/postcss-oklab-function/src/convert-oklch-to-srgb.ts
@@ -4,16 +4,21 @@ import { clip, inGamut, mapGamut } from './css-color-4/map-gamut';
 type color = [number, number, number];
 
 export function oklchToSRgb(oklchRaw: color): color {
-	const [lchLRaw, lchCRaw, lchHRaw] = oklchRaw;
+	const [oklchLRaw, oklchCRaw, oklchHRaw] = oklchRaw;
 
-	const lchL = Math.max(
-		lchLRaw,
+	const oklchL = Math.max(
+		oklchLRaw,
 		0,
 	);
 
-	const oklch = [lchL / 100, lchCRaw, lchHRaw % 360] as color;
+	const oklch = [oklchL / 100, oklchCRaw, oklchHRaw % 360] as color;
 
 	let conversion = oklch.slice() as color;
+	if (conversion[0] < 0.00001) { // very close to 0
+		// C and H components are powerless when L is 0 or very close to 0
+		conversion[1] = 0;
+		conversion[2] = 0;
+	}
 
 	conversion = OKLCH_to_OKLab(conversion);
 	conversion = OKLab_to_XYZ(conversion);